Visual Basic
A visual basic program I created in high school to display my understanding of menu strips. A number is input by the user and the program can display the absolute value, square root, tell whether it is positive or negative, as well as round the number to any number of places.